Extract test data injection into service
